+++ NEWS.md
@@ -1,5 +1,7 @@
 # watcher (development version)
 
+* Adds `$is_running()` method to the `Watcher` R6 class - use this to query the status as the `running` field is now private.
+
 # watcher 0.1.1
 
 * Updates bundled 'libfswatch' source package to 1.19.0-dev.